# openapi.json
{
  "components": {
    "schemas": {
...
      "createOrgReq": {
        "properties": {
          "name": {
            "type": "string"
          },
          "owner_email": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "orgView": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "memory_quota_bytes": {
            "type": "integer"
          },
          "monthly_cost_cap_usd": {
            "type": "number"
          },
          "monthly_run_quota": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "personal": {
            "type": "boolean"
          },
          "purge_after": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "suspend_reason": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ],
        "type": "object"
      },
      "setOrgStatusReq": {
        "properties": {
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "updateOrgReq": {
        "properties": {
          "memory_quota_bytes": {
            "type": "integer"
          },
          "monthly_cost_cap_usd": {
            "type": "number"
          },
          "monthly_run_quota": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Auto-generated route inventory for this iterion instance. Routes and methods are exact; request/response schemas are populated for the typed surface (see components) and enriched route-by-route. Note: the /api/v1/native, /api/v1/dispatcher and /api/v1/mcp/board sub-trees are served but registered on a separate mux, so they are intentionally not catalogued here.",
    "title": "iterion API",
    "version": "current"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/orgs": {
      "get": {
        "operationId": "getAdminOrgs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "orgs": {
                      "items": {
                        "$ref": "#/components/schemas/orgView"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "orgs"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "GET /api/admin/orgs",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "postAdminOrgs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createOrgReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "POST /api/admin/orgs",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orgs/{id}": {
      "delete": {
        "operationId": "deleteAdminOrgsById",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "DELETE /api/admin/orgs/{id}",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "getAdminOrgsById",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "GET /api/admin/orgs/{id}",
        "tags": [
          "admin"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "operationId": "patchAdminOrgsById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateOrgReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "PATCH /api/admin/orgs/{id}",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orgs/{id}/restore": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "postAdminOrgsByIdRestore",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "POST /api/admin/orgs/{id}/restore",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orgs/{id}/status": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "postAdminOrgsByIdStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/setOrgStatusReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orgView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "POST /api/admin/orgs/{id}/status",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orgs/{id}/teams": {
      "get": {
        "operationId": "getAdminOrgsByIdTeams",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/admin/orgs/{id}/teams",
        "tags": [
          "admin"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/admin/orgs/{id}/usage": {
      "get": {
        "operationId": "getAdminOrgsByIdUsage",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/admin/orgs/{id}/usage",
        "tags": [
          "admin"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}